publications: move PublicationForm to Views/
Components::PublicationForm -> Views::Controllers::Publications::Form
Per #4331. Also tightens body-wide assert_html label checks to
label[for='...'] (a body-text match would hit the first label, not
the field-specific one).
