+++ app/views/controllers/publications/edit.html.erb
@@ -4,4 +4,4 @@ add_edit_title(@publication)
 add_context_nav(publication_form_edit_tabs(pub: @publication))
 %>
 
-<%= render(Components::PublicationForm.new(@publication)) %>
+<%= render(Views::Controllers::Publications::Form.new(@publication)) %>
