+++ app/views/controllers/publications/new.html.erb
@@ -4,4 +4,4 @@ add_new_title(:create_object, :PUBLICATION)
 add_context_nav(publication_form_new_tabs)
 %>
 
-<%= render(Components::PublicationForm.new(@publication)) %>
+<%= render(Views::Controllers::Publications::Form.new(@publication)) %>
